fix : Kick MySelf - 어드민이 자기자신을 강퇴할 수 있던 현상 수정

+++ src/main/java/com/sparta/calendarprojectsnext/domain/user/service/UserService.java
@@ -73,6 +73,9 @@ public class UserService {
         if (!user.isAdmin()) {
             throw new CustomException(NOT_ADMIN);
         }
+        if (kickUserId.equals(user.getId())) {
+            throw new CustomException(NOT_KICK_SELF);
+        }
         User kickUser = userRepository.findById(kickUserId).orElseThrow(() -> new CustomException(USER_NOT_FOUND));
         userRepository.delete(kickUser);
     }
